sales_trading/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
celery
redis
psycopg2-binary
channels
channels-redis
drf-yasg
Pillow
reportlab